+++ example.js
@@ -1,7 +1,6 @@
-const fs = require('fs')
 const MBTiles = require('./')
 
-const mbtiles = new MBTiles('./test/in/plain_1.mbtiles')
-mbtiles.metadata().then(data => console.log(data))
-mbtiles.findOne([1, 2, 2]).then(data => fs.writeFileSync('tile.png', data))
-mbtiles.hashes().then(index => Object.keys(index))
+const mbtiles = new MBTiles('./test/out/blank2.mbtiles')
+// mbtiles.metadata().then(data => console.log(data))
+// mbtiles.findOne([1, 2, 2]).then(data => fs.writeFileSync('tile.png', data))
+mbtiles.hashes().then(index => console.log(index))